Round 26: fix 29 Core-layer review findings (function/IO/command edge correctness)
Full-FreeX round-26 review (16 deeper-core Sonnet lenses: meta regression-hunt on the round-25
merge + datetime/math-array/statistical-distribution/text-modern/error-array/io-comments-drawings/
io-pivot/calc-chain/cf-priority/table-structured-ref/form-controls/paste-operation/autofilter/
sheet-lifecycle; 2-of-2 verify) surfaced 37 findings (15 HIGH / 22 MED) — the HIGH count is
falling (25 -> 15) as the codebase converges. This change fixes 32 clean Core-layer findings
(each with a per-finding xUnit test plus a sibling no-regression test); 2 were rejected as
false positives; the ~3 host-UI findings are deferred.

Round-25 un-mirrored twin fixed:
- XLOOKUP's bare-range FAST path coerced an explicitly-supplied but blank if_not_found to #N/A
  (the round-25 fix only covered the slow path); now returns the supplied blank verbatim,
  matching the slow path and real Excel.

Core.Formula (function edges):
- 1900 phantom-leap-day (serial 60) family: DATE(1900,2,30)=61 / DATE(1900,2,31)=62 (was +1);
  YEARFRAC basis 1/2/3 and DAYS360 / YEARFRAC basis 0/4 now diff raw serials (or use day-29 for
  serial 60) instead of round-tripping through the collapsed DateTime — matching this file's own
  DAY(60)=29 and DATEDIF/DAYS serial-diff helpers.
- Distribution boundaries no longer return #NUM! from a 0*log(0)=NaN: BINOM.DIST(0,n,0)=1 and
  BINOM.DIST(n,n,1)=1 (degenerate mass); GAMMA.DIST pdf at x=0/alpha=1 = 1/beta; BETA.DIST pdf
  at the alpha=1 lower bound = uniform density.
- VALUETOTEXT/ARRAYTOTEXT format-0 vs format-1, TEXTSPLIT, a MAP/BYROW error-handling edge,
  FormulaSerializer structured-reference serialization, and a StructuredReferenceResolver edge.

Core.IO (round-trip fidelity):
- Legacy-comment preserver, OLE/control normalizer, pivot field read + pivot writer (NOT the
  deferred baseItem-index issue), slicer/timeline state, form-control property + linked-cell
  preserve, and a calcPr-adjacent writer fix.

Core.Commands / Core.Model / App.Services:
- Insert/Delete-Rows table/structured-ref shift, Paste Special Operation (Add/Sub/Mul/Div) +
  Skip-Blanks + Transpose combinations,
  sheet-lifecycle ref integrity (rename/delete/reorder/copy) in SheetCommands + Sheet.Clone +
  FormulaRewriter, and WorkbookOpenService calc-on-load (fullCalcOnLoad / trust-cache) correctness.

Rejected in triage (false positives, verified vs real Excel): R26-io-comments-drawings-deep-1
and R26-calc-chain-dependency-order-2.

Reverted in triage (regressed the intricate CF Manage-Rules flow across App.Presentation +
Avalonia — narrowing Applies-to and the priority Move-command): the 3 conditional-format
priority/apply-to findings (ApplyConditionalFormatCommand, PasteConditionalFormatsCommand,
ManageConditionalFormatsPlanner) are deferred for a dedicated CF-manage pass.

Deferred (host-UI): two MainWindow AutoFilter value-list/hidden-row host paths and a GridView
form-control input render finding.

Clean Release build + FreeX.DefaultTests green.

+++ src/FreeX.Core.Formula/BuiltInFunctions.TextSplit.cs
@@ -87,7 +87,11 @@ public static partial class BuiltInFunctions
             return TextResult("");
 
         var textLength = text.Length;
-        if (Math.Abs(options.InstanceNum) > textLength) return ErrorValue.Value;
+        // Use long arithmetic here: Math.Abs(int) throws OverflowException for int.MinValue
+        // (its positive counterpart isn't representable as an int), which would otherwise be
+        // swallowed by the generic OverflowException handler and surface as #NUM! instead of
+        // the #VALUE! every other instance_num domain violation in this function returns.
+        if (Math.Abs((long)options.InstanceNum) > textLength) return ErrorValue.Value;
 
         if (options.Delimiters.Any(delimiter => delimiter.Length == 0))
         {
